+++ lib/pages/signin_user_page/commands/signin_user_page_desktop_commands.dart
@@ -1,0 +1,5 @@
+import 'package:volume_vault/pages/signin_user_page/commands/signin_user_page_strategy.dart';
+
+class SigninUserPageDesktopCommands extends SigninUserPageStragegy {
+  //There is no specific methods for desktop.
+}
